Allow using exported types in component and instance types
This changes the syntax of component and instance types so that it is
possible to name a type with a typebound when it is declared, rather
than when it is exported, in order to create an entry for the abstract
type so created in the type index space which may be used in the types
of future exported definitions.  For uniformity's sake, the same
treatment is given to all definition sorts.